feat: implement critical frontend pages with zero-warning codebase
