+++ web/src/components/ui/badge.tsx
@@ -1,35 +1,8 @@
 import * as React from 'react'
-import { cva, type VariantProps } from 'class-variance-authority'
+import { type VariantProps } from 'class-variance-authority'
 
 import { cn } from '@/lib/utils'
-
-const badgeVariants = cva(
-  'inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2',
-  {
-    variants: {
-      variant: {
-        default:
-          'border-transparent bg-primary text-primary-foreground hover:bg-primary/80',
-        secondary:
-          'border-transparent bg-secondary text-secondary-foreground hover:bg-secondary/80',
-        destructive:
-          'border-transparent bg-destructive text-destructive-foreground hover:bg-destructive/80',
-        outline: 'text-foreground',
-        success:
-          'border-transparent bg-success text-white hover:bg-success/80',
-        warning:
-          'border-transparent bg-warning text-white hover:bg-warning/80',
-        error:
-          'border-transparent bg-error text-white hover:bg-error/80',
-        info:
-          'border-transparent bg-info text-white hover:bg-info/80',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-    },
-  }
-)
+import { badgeVariants } from '@/lib/badge-variants'
 
 export interface BadgeProps
   extends React.HTMLAttributes<HTMLDivElement>,
@@ -41,4 +14,4 @@ function Badge({ className, variant, ...props }: BadgeProps) {
   )
 }
 
-export { Badge, badgeVariants }
+export { Badge }
