+++ web/src/test/utils.tsx
@@ -95,5 +95,6 @@ export const mockApiResponses = {
 export const waitFor = (ms: number) => new Promise(resolve => setTimeout(resolve, ms))
 
 // Re-export testing utilities
+// eslint-disable-next-line react-refresh/only-export-components
 export * from '@testing-library/react'
 export { userEvent } from '@testing-library/user-event'
